rag_eval_project_v2/requirements.txt
beautifulsoup4>=4.12.3
chromadb>=0.5.20
dash>=2.18.2
dash-bootstrap-components>=1.6.0
duckduckgo-search>=6.3.5
firecrawl-py>=1.20.0
httpx>=0.28.1
llama-index-core>=0.11.23
llama-index-vector-stores-chroma>=0.2.1
numpy>=1.26.4,<2.0.0
openpyxl>=3.1.5
pandas>=2.2.3
plotly>=5.24.1
psutil>=7.2.2
python-dotenv>=1.0.1
pyyaml>=6.0.2
rank-bm25>=0.2.2
scikit-learn>=1.6.1
sentence-transformers>=5.4.1
streamlit>=1.44.1
tabulate>=0.9.0
tavily-python>=0.5.4
transformers==4.46.3
pytest>=8.3.5
pytest-asyncio>=0.26.0
